+++ src/examples/benchmarks/mul.ts
@@ -1,8 +1,10 @@
 /**
  * benchmark a circuit filled with generic gates
  */
-import { Circuit, Field, Provable, circuitMain, ZkProgram } from 'o1js';
+import { Experimental, Field, Provable, ZkProgram } from 'o1js';
 import { tic, toc } from '../utils/tic-toc.node.js';
+
+const { ZkFunction } = Experimental;
 
 // parameters
 let nMuls = (1 << 16) + (1 << 15); // not quite 2^17 generic gates = not quite 2^16 rows
@@ -25,13 +27,13 @@ async function getRows(nMuls: number) {
 }
 
 function simpleKimchiCircuit(nMuls: number) {
-  class MulChain extends Circuit {
-    @circuitMain
-    static run() {
+  return ZkFunction({
+    name: 'mul-chain',
+    privateInputTypes: [],
+    main: () => {
       main(nMuls);
-    }
-  }
-  return MulChain;
+    },
+  });
 }
 
 function picklesCircuit(nMuls: number) {
@@ -72,19 +74,19 @@ if (withPickles) {
   let circuit = simpleKimchiCircuit(nMuls);
 
   tic('compile 1 (includes srs creation)');
-  let kp = await circuit.generateKeypair();
+  let { verificationKey } = await circuit.compile();
   toc();
 
   tic('compile 2');
-  kp = await circuit.generateKeypair();
+  ({ verificationKey } = await circuit.compile());
   toc();
 
   tic('prove');
-  let p = await circuit.prove([], [], kp);
+  let p = await circuit.prove();
   toc();
 
   tic('verify');
-  let ok = await circuit.verify([], kp.verificationKey(), p);
+  let ok = await circuit.verify(p, verificationKey);
   toc();
   if (!ok) throw Error('invalid proof');
 }