+++ CHANGELOG.md
@@ -18,6 +18,17 @@ This project adheres to
 
 ## [Unreleased](https://github.com/o1-labs/o1js/compare/045b1ab...HEAD)
 
+### Deprecated
+
+- Deprecate the `Circuit` API in favor of the `ZkFunction` API.
+
+### Added
+
+- Added the `ZkFunction` API, exposed via the `Experimental` namespace
+  (`Experimental.ZkFunction`), which provides the same functionality as the
+  `Circuit` API with a more ergonomic and consistent structure, inspired by
+  `ZkProgram`. https://github.com/o1-labs/o1js/pull/2245
+
 ### Fixed
 
 - Sourcemaps properly work with the .cjs distribution of `o1js` now.